Two lint fixes.

+++ apropos_db.c
@@ -1,4 +1,4 @@
-/*	$Id: apropos_db.c,v 1.24 2011/12/16 12:06:35 kristaps Exp $ */
+/*	$Id: apropos_db.c,v 1.25 2011/12/16 12:07:45 kristaps Exp $ */
 /*
  * Copyright (c) 2011 Kristaps Dzonsons <kristaps@bsd.lv>
  * Copyright (c) 2011 Ingo Schwarze <schwarze@openbsd.org>
@@ -197,7 +197,7 @@ btree_read(const DBT *k, const DBT *v,
 static size_t
 norm_utf8(unsigned int cp, char out[7])
 {
-	size_t		 rc;
+	int		 rc;
 
 	rc = 0;
 
@@ -238,7 +238,7 @@ norm_utf8(unsigned int cp, char out[7])
 		return(0);
 
 	out[rc] = '\0';
-	return(rc);
+	return((size_t)rc);
 }
 
 /*
